\name{IlluminaHumanMethylation27kmanifest}
\alias{IlluminaHumanMethylation27kmanifest}
\docType{data}
\title{
  Manifest object for the IlluminaHumanMethylation27k array
}
\description{
  This object contains the array design for Illumina's Human Methylation
  27k microarray.
}
\usage{data(IlluminaHumanMethylation27kmanifest)}
\details{
  This package is based on the file
  \code{HumanMethylation27_270596_v1-2.csv}.
}
\format{
  An object of class \code{IlluminaMethylationManifest}
}
\seealso{
  The description of the class in the minfi package,
  \code{\link[minfi]{IlluminaMethylationManifest-class}}.
}
\examples{
data(IlluminaHumanMethylation27kmanifest)
}
\keyword{datasets}
